refactor: implement LocalSavableMovieDataSource with CRUD operations
